Transactioned chunking for dataset insertion
