+++ datasets/templates/datasets/dataset_validate.html
@@ -288,9 +288,11 @@
 		        .done(function(data) {
 		            if (data.status === "success") {
 		                const taskStatus = data.task_status;
+
+                        let stage = !!taskStatus.insert_start_time ? 'Ingesting' : 'Validating';
 		
 		                // Display file name and total features
-		                $('#file_name').html(`Validating: <b>${taskStatus.uploaded_filename}</b>: <i>${taskStatus.total_features} features</i>`);
+		                $('#file_name').html(`${stage}: <b>${taskStatus.uploaded_filename}</b>: <i>${taskStatus.total_features} features</i>`);
 		                
 		                if (!!taskStatus.remaining_queue) {
 		                	bar_text = `Waiting ... Job #${taskStatus.remaining_queue+1} in queue.`;
