+++ requirements.txt
@@ -141,3 +141,6 @@ vine==5.0.0
 wcwidth==0.2.6
 xlrd==2.0.1
 xlwt==1.3.0
+slack_sdk~=3.32.0
+python-dotenv~=1.0.1
+Jinja2~=3.1.4
